backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
faiss-cpu==1.8.0
pypdf==4.3.1
openai==1.50.0
python-dotenv==1.0.1
pydantic==2.9.2
tiktoken==0.7.0
